+++ Bizap/app/src/main/java/com/emul8r/bizap/domain/model/InvoiceSnapshot.kt
@@ -2,6 +2,7 @@ package com.emul8r.bizap.domain.model
 
 /**
  * IMMUTABLE SNAPSHOT OF INVOICE AT GENERATION TIME
+ * All monetary amounts are stored as Long (cents)
  */
 data class InvoiceSnapshot(
     val invoiceId: Long,
@@ -12,22 +13,22 @@ data class InvoiceSnapshot(
     val date: Long,
     val dueDate: Long,
     val items: List<LineItemSnapshot>,
-    val subtotal: Double,
+    val subtotal: Long,            // Cents
     val taxRate: Double,
-    val taxAmount: Double,
-    val totalAmount: Double,
+    val taxAmount: Long,           // Cents
+    val totalAmount: Long,         // Cents
     val businessName: String,
     val businessAbn: String,
     val businessEmail: String,
     val businessPhone: String,
     val businessAddress: String,
     val logoBase64: String?,
-    val currencyCode: String = "AUD" // NEW: Multi-currency support
+    val currencyCode: String = "AUD"
 )
 
 data class LineItemSnapshot(
     val description: String,
     val quantity: Double,
-    val unitPrice: Double,
-    val total: Double
+    val unitPrice: Long,           // Cents
+    val total: Long                // Cents (unitPrice * quantity)
 )
